Put TableDataConduit into its own file.
