+++ Makefile
@@ -140,7 +140,7 @@ docstest: install
 .PHONY: lint
 lint: install
 	@echo "Running ruff check ..."
-	@ruff check pixeltable/*.py pixeltable/functions pixeltable/index pixeltable/iterators pixeltable/metadata pixeltable/io/globals.py pixeltable/io/hf_datasets.py pixeltable/io/datarows.py pixeltable/io/pandas.py pixeltable/io/parquet.py pixeltable/io/utils.py pixeltable/io/external_store.py
+	@ruff check pixeltable/*.py pixeltable/functions pixeltable/index pixeltable/iterators pixeltable/metadata pixeltable/io/globals.py pixeltable/io/hf_datasets.py pixeltable/io/datarows.py pixeltable/io/pandas.py pixeltable/io/parquet.py pixeltable/io/utils.py pixeltable/io/external_store.py pixeltable/io/table_data_conduit.py
 
 .PHONY: formattest
 formattest: install
